Move expired tests to new suite
